Pivotal ID # 170993540: Refactor Processors To Services
Improve packages names
